fix bug where scheduling would fail when it should succeed
This bug happened when trying to schedule tasks on a timed-out
connection. As expected, the scheduleTask function initially failed to
queue the task and tried again. In the meantime however the task
parameter was replaced by its string representation. As a consequence
the second attempt received a string instead of an object, which
throws an exception when using json mode.